test(#315): integration tests for alerts API against ebull_test (ordering, clamp, dismiss-all scope)
Also fixes a real bug uncovered by the integration tests: the unseen_count query
used %(last_id)s IS NULL without a type cast, causing psycopg AmbiguousParameter
when last_id is None (PostgreSQL cannot infer BIGINT from an untyped NULL parameter).
Cast to ::BIGINT fixes it; unit test SQL-shape pin updated to match.

+++ app/api/alerts.py
@@ -92,7 +92,7 @@ def get_guard_rejections(
             WHERE pass_fail = 'FAIL'
               AND stage = 'execution_guard'
               AND decision_time >= now() - INTERVAL '7 days'
-              AND (%(last_id)s IS NULL OR decision_id > %(last_id)s)
+              AND (%(last_id)s::BIGINT IS NULL OR decision_id > %(last_id)s::BIGINT)
             """,
             {"last_id": last_seen},
         )
